Several changes in the same package:
- A top level Simulator class. Managues the Gui, the world, error system, etc.
- Start of the SaveXML feature. It is not complete, in fact it is not working right now.
I want to hear comments on this implementation. I will work on this feature the few days. When complete I'll add proper documentation to functions.
Only controller and one controller has been changed, it makes sense to add this functionality to controllers?
What about Joints?
- String class, allow us to have a simple SetValue () method instead of values for every variable type
The types are converted automagically thanks to the constructors
- A OgreCreator class.
I felt that OgreAdaptor was a miscellaneous class and creating entities was one of its tasks that could be isolated in a specific class for that. I am not sure is a great idea but at least some more rendering dependent code (CreatePlane, CreatBasicShapes) when to the rendering/ directory
- Start of the end of Global as it is now. The Gui is done for Gazebo on purpose, so we dont need to isolate it. Just to show the concept the Gui now call Simulator for exiting. In the near future Pausing time, counting time, etc. Will be moved to calls to Simulator. Other functionalities to the corresponding objects.

--HG--
branch : gazebo