add plugin path to LD_LIBRARY_PATH in setup.sh
Some ROS modules for gazebo link against some
of the more basic plugins gazebo provides.
This fails, however, if /usr/lib/gazebo*/plugins/
is not added to the library path.

This fix is present in the ubuntu package of gazebo,
but for some reason (lazyness?) not upstream.

--HG--
branch : setup.sh-ld-library-path