Adding explicit detection and linking of libavutil in libgazebo_common
Several functions from libavutil are used in libgazebo_common, but
only libavcodec was linked. While libavutil is a runtime libary dependency
of libavcodec, it does not become a direct runtime library dependency of
libgazebo_common. This means it will fail to link if LD_LIBRARY_PATH
is not set up properly or, more seriously, if using a linker which is
strict (like GNU GoLD).

tl;dr This enables Gazebo to be built with the GNU GoLD linker.

--HG--
branch : gazebo_2.2