Restore a minimum height of Gazebo window of ~700 pixels
The height of 703 pixels is obtained by moving the Stairs
button in the Building Editor .

Solves https://bitbucket.org/osrf/gazebo/issues/1706 .

Furthermore, add a warning when a user request in a gui.ini
file a window width or height that is below the minimum.

Add a tests that check that the minimum size of the window
is lower then some value (for the moment 700 pixels for the
width and 710 pixels to the height).

--HG--
branch : fixWindowMinSize