Respect solver iterations and preliminary GetParam() handling.
Needs work in SDF to support Siconos parameters propertly, for now we
steal "iters" from Bullet and treat it as the number of Newton
iterations, but this is not quite appropriate.

Default "iters" = 50, yet we usually use Newton iters=1 to 20, while
we need better control over solver iterations (usually 1000 or 10000),
solver tolerance (approx. 1e-4 to 1e-10), local solver iterations
(about 10 to 100), etc.